renames NewMenuItem files to MenuItem
